Get rid of non-generic data structures and switch to maphash.Comparable as MapOf hash function
