--- .github/workflows/build-1.18.yml
@@ -1,23 +1,0 @@
-name: build-1.18
-on: [push]
-
-jobs:
-  test:
-    runs-on: ubuntu-latest
-    strategy:
-      matrix:
-        go-version: [1.18.x]
-    name: Build with Go ${{ matrix.go-version }}
-    steps:
-      - uses: actions/checkout@v3
-
-      - name: Install Go
-        uses: actions/setup-go@v3
-        with:
-          go-version: ${{ matrix.go-version }}
-
-      - name: Run tests
-        run: go test -v ./...
-
-      - name: Run tests with race detector
-        run: go test -timeout 10m -race -v ./...
