+++ go.mod
@@ -1,3 +1,3 @@
-module github.com/puzpuzpuz/xsync/v3
+module github.com/puzpuzpuz/xsync/v4
 
-go 1.18
+go 1.24
