+++ mpmcqueueof_test.go
@@ -1,6 +1,3 @@
-//go:build go1.19
-// +build go1.19
-
 // Copyright notice. The following tests are partially based on
 // the following file from the Go Programming Language core repo:
 // https://github.com/golang/go/blob/831f9376d8d730b16fb33dfd775618dffe13ce7a/src/runtime/chan_test.go
@@ -13,9 +10,8 @@ import (
 	"sync"
 	"sync/atomic"
 	"testing"
-	"time"
 
-	. "github.com/puzpuzpuz/xsync/v3"
+	. "github.com/puzpuzpuz/xsync/v4"
 )
 
 func TestMPMCQueueOf_InvalidSize(t *testing.T) {
@@ -27,11 +23,13 @@ func TestMPMCQueueOf_InvalidSize(t *testing.T) {
 func TestMPMCQueueOfEnqueueDequeueInt(t *testing.T) {
 	q := NewMPMCQueueOf[int](10)
 	for i := 0; i < 10; i++ {
-		q.Enqueue(i)
+		if !q.TryEnqueue(i) {
+			t.Fatalf("failed to enqueue for %d", i)
+		}
 	}
 	for i := 0; i < 10; i++ {
-		if got := q.Dequeue(); got != i {
-			t.Fatalf("got %v, want %d", got, i)
+		if got, ok := q.TryDequeue(); !ok || got != i {
+			t.Fatalf("%v got %v, want %d", ok, got, i)
 		}
 	}
 }
@@ -39,11 +37,13 @@ func TestMPMCQueueOfEnqueueDequeueInt(t *testing.T) {
 func TestMPMCQueueOfEnqueueDequeueString(t *testing.T) {
 	q := NewMPMCQueueOf[string](10)
 	for i := 0; i < 10; i++ {
-		q.Enqueue(strconv.Itoa(i))
+		if !q.TryEnqueue(strconv.Itoa(i)) {
+			t.Fatalf("failed to enqueue for %d", i)
+		}
 	}
 	for i := 0; i < 10; i++ {
-		if got := q.Dequeue(); got != strconv.Itoa(i) {
-			t.Fatalf("got %v, want %d", got, i)
+		if got, ok := q.TryDequeue(); !ok || got != strconv.Itoa(i) {
+			t.Fatalf("%v got %v, want %d", ok, got, i)
 		}
 	}
 }
@@ -55,11 +55,13 @@ func TestMPMCQueueOfEnqueueDequeueStruct(t *testing.T) {
 	}
 	q := NewMPMCQueueOf[foo](10)
 	for i := 0; i < 10; i++ {
-		q.Enqueue(foo{i, i})
+		if !q.TryEnqueue(foo{i, i}) {
+			t.Fatalf("failed to enqueue for %d", i)
+		}
 	}
 	for i := 0; i < 10; i++ {
-		if got := q.Dequeue(); got.bar != i || got.baz != i {
-			t.Fatalf("got %v, want %d", got, i)
+		if got, ok := q.TryDequeue(); !ok || got.bar != i || got.baz != i {
+			t.Fatalf("%v got %v, want %d", ok, got, i)
 		}
 	}
 }
@@ -71,54 +73,21 @@ func TestMPMCQueueOfEnqueueDequeueStructRef(t *testing.T) {
 	}
 	q := NewMPMCQueueOf[*foo](11)
 	for i := 0; i < 10; i++ {
-		q.Enqueue(&foo{i, i})
+		if !q.TryEnqueue(&foo{i, i}) {
+			t.Fatalf("failed to enqueue for %d", i)
+		}
 	}
-	q.Enqueue(nil)
+	if !q.TryEnqueue(nil) {
+		t.Fatal("failed to enqueue for nil")
+	}
 	for i := 0; i < 10; i++ {
-		if got := q.Dequeue(); got.bar != i || got.baz != i {
-			t.Fatalf("got %v, want %d", got, i)
+		if got, ok := q.TryDequeue(); !ok || got.bar != i || got.baz != i {
+			t.Fatalf("%v got %v, want %d", ok, got, i)
 		}
 	}
-	if last := q.Dequeue(); last != nil {
-		t.Fatalf("got %v, want nil", last)
+	if last, ok := q.TryDequeue(); !ok || last != nil {
+		t.Fatalf("%v got %v, want nil", ok, last)
 	}
-}
-
-func TestMPMCQueueOfEnqueueBlocksOnFull(t *testing.T) {
-	q := NewMPMCQueueOf[string](1)
-	q.Enqueue("foo")
-	cdone := make(chan bool)
-	flag := int32(0)
-	go func() {
-		q.Enqueue("bar")
-		if atomic.LoadInt32(&flag) == 0 {
-			t.Error("enqueue on full queue didn't wait for dequeue")
-		}
-		cdone <- true
-	}()
-	time.Sleep(50 * time.Millisecond)
-	atomic.StoreInt32(&flag, 1)
-	if got := q.Dequeue(); got != "foo" {
-		t.Fatalf("got %v, want foo", got)
-	}
-	<-cdone
-}
-
-func TestMPMCQueueOfDequeueBlocksOnEmpty(t *testing.T) {
-	q := NewMPMCQueueOf[string](2)
-	cdone := make(chan bool)
-	flag := int32(0)
-	go func() {
-		q.Dequeue()
-		if atomic.LoadInt32(&flag) == 0 {
-			t.Error("dequeue on empty queue didn't wait for enqueue")
-		}
-		cdone <- true
-	}()
-	time.Sleep(50 * time.Millisecond)
-	atomic.StoreInt32(&flag, 1)
-	q.Enqueue("foobar")
-	<-cdone
 }
 
 func TestMPMCQueueOfTryEnqueueDequeue(t *testing.T) {
@@ -150,62 +119,6 @@ func TestMPMCQueueOfTryDequeueOnEmpty(t *testing.T) {
 	if _, ok := q.TryDequeue(); ok {
 		t.Error("got success for enqueue on empty queue")
 	}
-}
-
-func hammerMPMCQueueOfBlockingCalls(t *testing.T, gomaxprocs, numOps, numThreads int) {
-	runtime.GOMAXPROCS(gomaxprocs)
-	q := NewMPMCQueueOf[int](numThreads)
-	startwg := sync.WaitGroup{}
-	startwg.Add(1)
-	csum := make(chan int, numThreads)
-	// Start producers.
-	for i := 0; i < numThreads; i++ {
-		go func(n int) {
-			startwg.Wait()
-			for j := n; j < numOps; j += numThreads {
-				q.Enqueue(j)
-			}
-		}(i)
-	}
-	// Start consumers.
-	for i := 0; i < numThreads; i++ {
-		go func(n int) {
-			startwg.Wait()
-			sum := 0
-			for j := n; j < numOps; j += numThreads {
-				item := q.Dequeue()
-				sum += item
-			}
-			csum <- sum
-		}(i)
-	}
-	startwg.Done()
-	// Wait for all the sums from producers.
-	sum := 0
-	for i := 0; i < numThreads; i++ {
-		s := <-csum
-		sum += s
-	}
-	// Assert the total sum.
-	expectedSum := numOps * (numOps - 1) / 2
-	if sum != expectedSum {
-		t.Fatalf("sums don't match for %d num ops, %d num threads: got %d, want %d",
-			numOps, numThreads, sum, expectedSum)
-	}
-}
-
-func TestMPMCQueueOfBlockingCalls(t *testing.T) {
-	defer runtime.GOMAXPROCS(runtime.GOMAXPROCS(-1))
-	n := 100
-	if testing.Short() {
-		n = 10
-	}
-	hammerMPMCQueueOfBlockingCalls(t, 1, 100*n, n)
-	hammerMPMCQueueOfBlockingCalls(t, 1, 1000*n, 10*n)
-	hammerMPMCQueueOfBlockingCalls(t, 4, 100*n, n)
-	hammerMPMCQueueOfBlockingCalls(t, 4, 1000*n, 10*n)
-	hammerMPMCQueueOfBlockingCalls(t, 8, 100*n, n)
-	hammerMPMCQueueOfBlockingCalls(t, 8, 1000*n, 10*n)
 }
 
 func hammerMPMCQueueOfNonBlockingCalls(t *testing.T, gomaxprocs, numOps, numThreads int) {
@@ -290,16 +203,30 @@ func benchmarkMPMCQueueOf(b *testing.B, queueSize, localWork int) {
 						foo *= 2
 						foo /= 2
 					}
-					q.Enqueue(1)
+					for !q.TryEnqueue(1) {
+						runtime.Gosched()
+					}
 				}
 			}
-			q.Enqueue(0)
+			for !q.TryEnqueue(0) {
+				runtime.Gosched()
+			}
 			c <- foo == 42
 		}()
 		go func() {
 			foo := 0
 			for {
-				v := q.Dequeue()
+				var (
+					v  int
+					ok bool
+				)
+				for {
+					if v, ok = q.TryDequeue(); !ok {
+						runtime.Gosched()
+					} else {
+						break
+					}
+				}
 				if v == 0 {
 					break
 				}
