+++ spscqueueof.go
@@ -1,6 +1,3 @@
-//go:build go1.19
-// +build go1.19
-
 package xsync
 
 import (
